tests running with refactor complete.  Need to remove new code still
