+++ bcda/cclf/cclf_test.go
@@ -1,117 +1,38 @@
 package cclf
 
 import (
-	"bytes"
-	"fmt"
+	"github.com/CMSgov/bcda-app/bcda/testUtils"
 	"io/ioutil"
 	"os"
 	"path/filepath"
-	"strings"
 	"testing"
 	"time"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/suite"
-	"github.com/urfave/cli"
 
 	"github.com/CMSgov/bcda-app/bcda/database"
 	"github.com/CMSgov/bcda-app/bcda/models"
 )
 
-type CLITestSuite struct {
+const BASE_FILE_PATH = "../../shared_files/"
+
+type CCLFTestSuite struct {
 	suite.Suite
-	testApp *cli.App
 }
 
-func (s *CLITestSuite) SetupTest() {
-	s.testApp = setUpApp()
+func (s *CCLFTestSuite) SetupTest() {
 	models.InitializeGormModels()
 }
 
-func TestCLITestSuite(t *testing.T) {
-	suite.Run(t, new(CLITestSuite))
+func TestCCLFTestSuite(t *testing.T) {
+	suite.Run(t, new(CCLFTestSuite))
 }
 
-func (s *CLITestSuite) TestCreateACO() {
-
-	// init
-	db := database.GetGORMDbConnection()
-	defer database.Close(db)
-
-	// set up the test app writer (to redirect CLI responses from stdout to a byte buffer)
-	buf := new(bytes.Buffer)
-	s.testApp.Writer = buf
-
+func (s *CCLFTestSuite) TestImportCCLF0() {
 	assert := assert.New(s.T())
 
-	// Successful ACO creation
-	ACOName := "Unit Test ACO 1"
-	args := []string{"bcda", "create-aco", "--name", ACOName}
-	err := s.testApp.Run(args)
-	assert.Nil(err)
-	assert.NotNil(buf)
-	acoUUID := strings.TrimSpace(buf.String())
-	var testACO models.ACO
-	db.First(&testACO, "Name=?", ACOName)
-	assert.Equal(testACO.UUID.String(), acoUUID)
-	buf.Reset()
-
-	ACO2Name := "Unit Test ACO 2"
-	aco2ID := "A9999"
-	args = []string{"bcda", "create-aco", "--name", ACO2Name, "--cms-id", aco2ID}
-	err = s.testApp.Run(args)
-	assert.Nil(err)
-	assert.NotNil(buf)
-	acoUUID = strings.TrimSpace(buf.String())
-	var testACO2 models.ACO
-	db.First(&testACO2, "Name=?", ACO2Name)
-	assert.Equal(testACO2.UUID.String(), acoUUID)
-	assert.Equal(*testACO2.CMSID, aco2ID)
-	buf.Reset()
-
-	// Negative tests
-
-	// No parameters
-	args = []string{"bcda", "create-aco"}
-	err = s.testApp.Run(args)
-	assert.Equal("ACO name (--name) must be provided", err.Error())
-	assert.Equal(0, buf.Len())
-	buf.Reset()
-
-	// No ACO Name
-	badACO := ""
-	args = []string{"bcda", "create-aco", "--name", badACO}
-	err = s.testApp.Run(args)
-	assert.Equal("ACO name (--name) must be provided", err.Error())
-	assert.Equal(0, buf.Len())
-	buf.Reset()
-
-	// ACO name without flag
-	args = []string{"bcda", "create-aco", ACOName}
-	err = s.testApp.Run(args)
-	assert.Equal("ACO name (--name) must be provided", err.Error())
-	assert.Equal(0, buf.Len())
-	buf.Reset()
-
-	// Unexpected flag
-	args = []string{"bcda", "create-aco", "--abcd", "efg"}
-	err = s.testApp.Run(args)
-	assert.Equal("flag provided but not defined: -abcd", err.Error())
-	assert.Contains(buf.String(), "Incorrect Usage: flag provided but not defined")
-	buf.Reset()
-
-	// Invalid CMS ID
-	args = []string{"bcda", "create-aco", "--name", ACOName, "--cms-id", "ABCDE"}
-	err = s.testApp.Run(args)
-	assert.Equal("ACO CMS ID (--cms-id) is invalid", err.Error())
-	assert.Equal(0, buf.Len())
-	buf.Reset()
-}
-
-func (s *CLITestSuite) TestImportCCLF0() {
-	assert := assert.New(s.T())
-
-	cclf0filePath := "../shared_files/cclf/T.A0001.ACO.ZC0Y18.D181120.T1000011"
+	cclf0filePath := BASE_FILE_PATH + "cclf/T.A0001.ACO.ZC0Y18.D181120.T1000011"
 	cclf0metadata := &cclfFileMetadata{env: "test", acoID: "A0001", cclfNum: 0, timestamp: time.Now(), filePath: cclf0filePath, perfYear: 18}
 
 	// positive
@@ -126,21 +47,21 @@ func (s *CLITestSuite) TestImportCCLF0() {
 	assert.NotNil(err)
 
 	// missing cclf8 and or 9 from cclf0
-	cclf0filePath = "../shared_files/cclf0_MissingData/T.A0001.ACO.ZC0Y18.D181120.T1000011"
+	cclf0filePath = BASE_FILE_PATH + "cclf0_MissingData/T.A0001.ACO.ZC0Y18.D181120.T1000011"
 	cclf0metadata = &cclfFileMetadata{env: "test", acoID: "A0001", cclfNum: 0, timestamp: time.Now(), filePath: cclf0filePath, perfYear: 18}
 	_, err = importCCLF0(cclf0metadata)
 	assert.NotNil(err)
 
-	cclf0filePath = "../shared_files/cclf0_MissingData/T.A0001.ACO.ZC0Y18.D181120.T1000012"
+	cclf0filePath = BASE_FILE_PATH + "cclf0_MissingData/T.A0001.ACO.ZC0Y18.D181120.T1000012"
 	cclf0metadata = &cclfFileMetadata{env: "test", acoID: "A0001", cclfNum: 0, timestamp: time.Now(), filePath: cclf0filePath, perfYear: 18}
 	_, err = importCCLF0(cclf0metadata)
 	assert.NotNil(err)
 }
 
-func (s *CLITestSuite) TestImportCCLF0_SplitFiles() {
+func (s *CCLFTestSuite) TestImportCCLF0_SplitFiles() {
 	assert := assert.New(s.T())
 
-	cclf0filePath := "../shared_files/cclf_split/T.A0001.ACO.ZC0Y18.D181120.T1000011"
+	cclf0filePath := BASE_FILE_PATH + "cclf_split/T.A0001.ACO.ZC0Y18.D181120.T1000011"
 	cclf0metadata := &cclfFileMetadata{env: "test", acoID: "A0001", cclfNum: 0, timestamp: time.Now(), filePath: cclf0filePath, perfYear: 18}
 
 	validator, err := importCCLF0(cclf0metadata)
@@ -149,13 +70,13 @@ func (s *CLITestSuite) TestImportCCLF0_SplitFiles() {
 	assert.Equal(cclfFileValidator{totalRecordCount: 6, maxRecordLength: 54}, validator["CCLF9"])
 }
 
-func (s *CLITestSuite) TestValidate() {
+func (s *CCLFTestSuite) TestValidate() {
 	assert := assert.New(s.T())
 
-	cclf8filePath := "../shared_files/cclf/T.A0001.ACO.ZC8Y18.D181120.T1000009"
+	cclf8filePath := BASE_FILE_PATH + "cclf/T.A0001.ACO.ZC8Y18.D181120.T1000009"
 	cclf8metadata := &cclfFileMetadata{env: "test", acoID: "A0001", cclfNum: 8, timestamp: time.Now(), filePath: cclf8filePath, perfYear: 18}
 
-	cclf9filePath := "../shared_files/cclf/T.A0001.ACO.ZC9Y18.D181120.T1000010"
+	cclf9filePath := BASE_FILE_PATH + "cclf/T.A0001.ACO.ZC9Y18.D181120.T1000010"
 	cclf9metadata := &cclfFileMetadata{env: "test", acoID: "A0001", cclfNum: 9, timestamp: time.Now(), filePath: cclf9filePath, perfYear: 18}
 
 	// positive
@@ -173,14 +94,14 @@ func (s *CLITestSuite) TestValidate() {
 	assert.NotNil(err)
 }
 
-func (s *CLITestSuite) TestValidate_SplitFiles() {
+func (s *CCLFTestSuite) TestValidate_SplitFiles() {
 	assert := assert.New(s.T())
 
 	cclf8Metadata := &cclfFileMetadata{
 		acoID:     "A0001",
 		cclfNum:   8,
 		timestamp: time.Now(),
-		filePath:  "../shared_files/cclf_split/T.A0001.ACO.ZC8Y18.D181120.T1000009",
+		filePath:  BASE_FILE_PATH + "cclf_split/T.A0001.ACO.ZC8Y18.D181120.T1000009",
 		perfYear:  18,
 	}
 
@@ -188,7 +109,7 @@ func (s *CLITestSuite) TestValidate_SplitFiles() {
 		acoID:     "A0001",
 		cclfNum:   8,
 		timestamp: time.Now(),
-		filePath:  "../shared_files/cclf_split/T.A0001.ACO.ZC9Y18.D181120.T1000010",
+		filePath:  BASE_FILE_PATH + "cclf_split/T.A0001.ACO.ZC9Y18.D181120.T1000010",
 		perfYear:  18,
 	}
 
@@ -204,7 +125,7 @@ func (s *CLITestSuite) TestValidate_SplitFiles() {
 	assert.Nil(err)
 }
 
-func (s *CLITestSuite) TestImportCCLF8() {
+func (s *CCLFTestSuite) TestImportCCLF8() {
 	assert := assert.New(s.T())
 	db := database.GetGORMDbConnection()
 	defer database.Close(db)
@@ -221,7 +142,7 @@ func (s *CLITestSuite) TestImportCCLF8() {
 		cclfNum:   8,
 		perfYear:  18,
 		timestamp: fileTime,
-		filePath:  "../shared_files/cclf/T.A0001.ACO.ZC8Y18.D181120.T1000009",
+		filePath:  BASE_FILE_PATH + "cclf/T.A0001.ACO.ZC8Y18.D181120.T1000009",
 	}
 
 	err := importCCLF8(metadata)
@@ -257,7 +178,7 @@ func (s *CLITestSuite) TestImportCCLF8() {
 	db.Unscoped().Delete(&models.CCLFFile{})
 }
 
-func (s *CLITestSuite) TestImportCCLF8_SplitFiles() {
+func (s *CCLFTestSuite) TestImportCCLF8_SplitFiles() {
 	assert := assert.New(s.T())
 	db := database.GetGORMDbConnection()
 	defer database.Close(db)
@@ -274,7 +195,7 @@ func (s *CLITestSuite) TestImportCCLF8_SplitFiles() {
 		cclfNum:   8,
 		perfYear:  18,
 		timestamp: fileTime,
-		filePath:  "../shared_files/cclf_split/T.A0001.ACO.ZC8Y18.D181120.T1000009",
+		filePath:  BASE_FILE_PATH + "cclf_split/T.A0001.ACO.ZC8Y18.D181120.T1000009",
 	}
 
 	err := importCCLF8(metadata)
@@ -310,7 +231,7 @@ func (s *CLITestSuite) TestImportCCLF8_SplitFiles() {
 	db.Unscoped().Delete(&models.CCLFFile{})
 }
 
-func (s *CLITestSuite) TestImportCCLF8_InvalidMetadata() {
+func (s *CCLFTestSuite) TestImportCCLF8_InvalidMetadata() {
 	assert := assert.New(s.T())
 
 	var metadata *cclfFileMetadata
@@ -319,7 +240,7 @@ func (s *CLITestSuite) TestImportCCLF8_InvalidMetadata() {
 	assert.EqualError(err, "CCLF file not found")
 }
 
-func (s *CLITestSuite) TestImportCCLF9() {
+func (s *CCLFTestSuite) TestImportCCLF9() {
 	assert := assert.New(s.T())
 	db := database.GetGORMDbConnection()
 	defer database.Close(db)
@@ -334,7 +255,7 @@ func (s *CLITestSuite) TestImportCCLF9() {
 		acoID:     acoID,
 		cclfNum:   9,
 		timestamp: fileTime,
-		filePath:  "../shared_files/cclf/T.A0001.ACO.ZC9Y18.D181120.T1000010",
+		filePath:  BASE_FILE_PATH + "cclf/T.A0001.ACO.ZC9Y18.D181120.T1000010",
 		perfYear:  18,
 		name:      "T.A0001.ACO.ZC9Y18.D181120.T1000010",
 	}
@@ -363,7 +284,7 @@ func (s *CLITestSuite) TestImportCCLF9() {
 	db.Unscoped().Delete(&models.CCLFFile{})
 }
 
-func (s *CLITestSuite) TestImportCCLF9_SplitFiles() {
+func (s *CCLFTestSuite) TestImportCCLF9_SplitFiles() {
 	assert := assert.New(s.T())
 	db := database.GetGORMDbConnection()
 	defer database.Close(db)
@@ -378,7 +299,7 @@ func (s *CLITestSuite) TestImportCCLF9_SplitFiles() {
 		acoID:     acoID,
 		cclfNum:   9,
 		timestamp: fileTime,
-		filePath:  "../shared_files/cclf_split/T.A0001.ACO.ZC9Y18.D181120.T1000010",
+		filePath:  BASE_FILE_PATH + "cclf_split/T.A0001.ACO.ZC9Y18.D181120.T1000010",
 		perfYear:  18,
 		name:      "T.A0001.ACO.ZC9Y18.D181120.T1000010",
 	}
@@ -407,7 +328,7 @@ func (s *CLITestSuite) TestImportCCLF9_SplitFiles() {
 	db.Unscoped().Delete(&models.CCLFFile{})
 }
 
-func (s *CLITestSuite) TestImportCCLF9_InvalidMetadata() {
+func (s *CCLFTestSuite) TestImportCCLF9_InvalidMetadata() {
 	assert := assert.New(s.T())
 
 	var metadata *cclfFileMetadata
@@ -416,7 +337,7 @@ func (s *CLITestSuite) TestImportCCLF9_InvalidMetadata() {
 	assert.EqualError(err, "CCLF file not found")
 }
 
-func (s *CLITestSuite) TestGetCCLFFileMetadata() {
+func (s *CCLFTestSuite) TestGetCCLFFileMetadata() {
 	assert := assert.New(s.T())
 
 	_, err := getCCLFFileMetadata("/path/to/file")
@@ -459,12 +380,12 @@ func (s *CLITestSuite) TestGetCCLFFileMetadata() {
 	assert.NotNil(err)
 }
 
-func (s *CLITestSuite) TestSortCCLFFiles() {
+func (s *CCLFTestSuite) TestSortCCLFFiles() {
 	assert := assert.New(s.T())
 	cclfmap := make(map[string][]*cclfFileMetadata)
 	var skipped int
 
-	filePath := "../shared_files/cclf/"
+	filePath := BASE_FILE_PATH + "cclf/"
 	err := filepath.Walk(filePath, sortCCLFFiles(&cclfmap, &skipped))
 	assert.Nil(err)
 	assert.Equal(3, len(cclfmap["A0001_18"]))
@@ -472,7 +393,7 @@ func (s *CLITestSuite) TestSortCCLFFiles() {
 
 	cclfmap = make(map[string][]*cclfFileMetadata)
 	skipped = 0
-	filePath = "../shared_files/cclf_BadFileNames/"
+	filePath = BASE_FILE_PATH + "cclf_BadFileNames/"
 	err = filepath.Walk(filePath, sortCCLFFiles(&cclfmap, &skipped))
 	assert.Nil(err)
 	cclflist := cclfmap["A0001_18"]
@@ -481,11 +402,11 @@ func (s *CLITestSuite) TestSortCCLFFiles() {
 	for _, cclf := range cclflist {
 		assert.NotEqual(s.T(), 9, cclf.cclfNum)
 	}
-	s.resetFiles("../shared_files/cclf_BadFileNames/")
+	testUtils.ResetFiles(s.Suite, BASE_FILE_PATH+"cclf_BadFileNames/")
 
 	cclfmap = make(map[string][]*cclfFileMetadata)
 	skipped = 0
-	filePath = "../shared_files/cclf0/"
+	filePath = BASE_FILE_PATH + "cclf0/"
 	err = filepath.Walk(filePath, sortCCLFFiles(&cclfmap, &skipped))
 	assert.Nil(err)
 	cclflist = cclfmap["A0001_18"]
@@ -497,7 +418,7 @@ func (s *CLITestSuite) TestSortCCLFFiles() {
 
 	cclfmap = make(map[string][]*cclfFileMetadata)
 	skipped = 0
-	filePath = "../shared_files/cclf8/"
+	filePath = BASE_FILE_PATH + "cclf8/"
 	err = filepath.Walk(filePath, sortCCLFFiles(&cclfmap, &skipped))
 	assert.Nil(err)
 	cclflist = cclfmap["A0001_18"]
@@ -509,7 +430,7 @@ func (s *CLITestSuite) TestSortCCLFFiles() {
 
 	cclfmap = make(map[string][]*cclfFileMetadata)
 	skipped = 0
-	filePath = "../shared_files/cclf9/"
+	filePath = BASE_FILE_PATH + "cclf9/"
 	err = filepath.Walk(filePath, sortCCLFFiles(&cclfmap, &skipped))
 	assert.Nil(err)
 	cclflist = cclfmap["A0001_18"]
@@ -529,7 +450,7 @@ func (s *CLITestSuite) TestSortCCLFFiles() {
 	}
 
 	cclfmap = make(map[string][]*cclfFileMetadata)
-	filePath = "../shared_files/cclf9/"
+	filePath = BASE_FILE_PATH + "cclf9/"
 	err = filepath.Walk(filePath, sortCCLFFiles(&cclfmap, &skipped))
 	assert.Nil(err)
 	cclflist = cclfmap["A0001_18"]
@@ -540,7 +461,7 @@ func (s *CLITestSuite) TestSortCCLFFiles() {
 
 	cclfmap = make(map[string][]*cclfFileMetadata)
 	skipped = 0
-	filePath = "../shared_files/cclf_All/"
+	filePath = BASE_FILE_PATH + "cclf_All/"
 	err = filepath.Walk(filePath, sortCCLFFiles(&cclfmap, &skipped))
 	assert.Nil(err)
 	cclflist = cclfmap["A0001_18"]
@@ -562,49 +483,9 @@ func (s *CLITestSuite) TestSortCCLFFiles() {
 	assert.Equal(3, len(cclf0))
 }
 
-func (s *CLITestSuite) TestImportCCLFDirectory() {
-	assert := assert.New(s.T())
-
-	db := database.GetGORMDbConnection()
-	defer database.Close(db)
-
-	db.Unscoped().Delete(&models.CCLFBeneficiary{})
-	db.Unscoped().Delete(&models.CCLFFile{})
-
-	// set up the test app writer (to redirect CLI responses from stdout to a byte buffer)
-	buf := new(bytes.Buffer)
-	s.testApp.Writer = buf
-
-	s.setPendingDeletionDir()
-
-	args := []string{"bcda", "import-cclf-directory", "--directory", "../shared_files/cclf/"}
-	err := s.testApp.Run(args)
-	assert.Nil(err)
-	assert.Contains(buf.String(), "Completed CCLF import.")
-	assert.Contains(buf.String(), "Successfully imported 3 files.")
-	assert.Contains(buf.String(), "Failed to import 0 files.")
-	assert.Contains(buf.String(), "Skipped 0 files.")
-
-	buf.Reset()
-	db.Unscoped().Delete(&models.CCLFBeneficiary{})
-	db.Unscoped().Delete(&models.CCLFFile{})
-	s.resetFiles("../shared_files/cclf/")
-
-	// dir has 4 files, but 2 will be ignored because of bad file names.
-	args = []string{"bcda", "import-cclf-directory", "--directory", "../shared_files/cclf_BadFileNames/"}
-	err = s.testApp.Run(args)
-	assert.NotNil(err)
-	assert.Contains(buf.String(), "Completed CCLF import.")
-	assert.Contains(buf.String(), "Successfully imported 2 files.")
-	assert.Contains(buf.String(), "Failed to import 1 files.")
-	assert.Contains(buf.String(), "Skipped 2 files.")
-	buf.Reset()
-
-	s.resetFiles("../shared_files/cclf_BadFileNames/")
-}
-func (s *CLITestSuite) TestCleanupCCLF() {
+func (s *CCLFTestSuite) TestCleanupCCLF() {
 	cclfmap := make(map[string][]*cclfFileMetadata)
-	s.setPendingDeletionDir()
+	testUtils.SetPendingDeletionDir(s.Suite)
 
 	// failed import: file that's within the threshold - stay put
 	acoID := "A0001"
@@ -616,7 +497,7 @@ func (s *CLITestSuite) TestCleanupCCLF() {
 		cclfNum:      8,
 		perfYear:     18,
 		timestamp:    fileTime,
-		filePath:     "../shared_files/cclf/T.A0001.ACO.ZC0Y18.D181120.T1000011",
+		filePath:     BASE_FILE_PATH + "cclf/T.A0001.ACO.ZC0Y18.D181120.T1000011",
 		imported:     false,
 		deliveryDate: time.Now(),
 	}
@@ -631,7 +512,7 @@ func (s *CLITestSuite) TestCleanupCCLF() {
 		cclfNum:      8,
 		perfYear:     18,
 		timestamp:    fileTime,
-		filePath:     "../shared_files/cclf/T.A0001.ACO.ZC8Y18.D181120.T1000009",
+		filePath:     BASE_FILE_PATH + "cclf/T.A0001.ACO.ZC8Y18.D181120.T1000009",
 		imported:     false,
 		deliveryDate: fileTime,
 	}
@@ -646,7 +527,7 @@ func (s *CLITestSuite) TestCleanupCCLF() {
 		cclfNum:   9,
 		perfYear:  18,
 		timestamp: fileTime,
-		filePath:  "../shared_files/cclf/T.A0001.ACO.ZC9Y18.D181120.T1000010",
+		filePath:  BASE_FILE_PATH + "cclf/T.A0001.ACO.ZC9Y18.D181120.T1000010",
 		imported:  true,
 	}
 	cclfmap["A0001_18"] = []*cclfFileMetadata{cclf0metadata, cclf8metadata, cclf9metadata}
@@ -659,36 +540,13 @@ func (s *CLITestSuite) TestCleanupCCLF() {
 	for _, file := range files {
 		assert.NotEqual(s.T(), "T.A0001.ACO.ZC0Y18.D181120.T1000011", file.Name())
 	}
-	s.resetFiles("../shared_files/cclf/")
+	testUtils.ResetFiles(s.Suite, BASE_FILE_PATH+"cclf/")
 }
 
-func (s *CLITestSuite) TestImportCCLFDirectory_SplitFiles() {
+func (s *CCLFTestSuite) TestDeleteDirectory() {
 	assert := assert.New(s.T())
-
-	db := database.GetGORMDbConnection()
-	defer database.Close(db)
-
-	db.Unscoped().Delete(&models.CCLFBeneficiary{})
-	db.Unscoped().Delete(&models.CCLFFile{})
-
-	buf := new(bytes.Buffer)
-	s.testApp.Writer = buf
-
-	args := []string{"bcda", "import-cclf-directory", "--directory", "../shared_files/cclf_split/"}
-	err := s.testApp.Run(args)
-	assert.Nil(err)
-	assert.Contains(buf.String(), "Completed CCLF import.")
-	assert.Contains(buf.String(), "Successfully imported 3 files.")
-	assert.Contains(buf.String(), "Failed to import 0 files.")
-	assert.Contains(buf.String(), "Skipped 0 files.")
-
-	s.resetFiles("../shared_files/cclf_split/")
-}
-
-func (s *CLITestSuite) TestDeleteDirectory() {
-	assert := assert.New(s.T())
-	dirToDelete := "../shared_files/doomedDirectory"
-	s.makeDirToDelete(dirToDelete)
+	dirToDelete := BASE_FILE_PATH + "doomedDirectory"
+	testUtils.MakeDirToDelete(s.Suite, dirToDelete)
 	defer os.Remove(dirToDelete)
 
 	f, err := os.Open(dirToDelete)
@@ -697,7 +555,7 @@ func (s *CLITestSuite) TestDeleteDirectory() {
 	assert.Nil(err)
 	assert.Equal(4, len(files))
 
-	filesDeleted, err := deleteDirectoryContents(dirToDelete)
+	filesDeleted, err := DeleteDirectoryContents(dirToDelete)
 	assert.Equal(4, filesDeleted)
 	assert.Nil(err)
 
@@ -707,87 +565,7 @@ func (s *CLITestSuite) TestDeleteDirectory() {
 	assert.Nil(err)
 	assert.Equal(0, len(files))
 
-	filesDeleted, err = deleteDirectoryContents("This/Does/not/Exist")
+	filesDeleted, err = DeleteDirectoryContents("This/Does/not/Exist")
 	assert.Equal(0, filesDeleted)
 	assert.NotNil(err)
-}
-
-func (s *CLITestSuite) TestDeleteDirectoryContents() {
-	assert := assert.New(s.T())
-	buf := new(bytes.Buffer)
-	s.testApp.Writer = buf
-
-	dirToDelete := "../shared_files/doomedDirectory"
-	s.makeDirToDelete(dirToDelete)
-	defer os.Remove(dirToDelete)
-
-	args := []string{"bcda", "delete-dir-contents", "--dirToDelete", dirToDelete}
-	err := s.testApp.Run(args)
-	assert.Nil(err)
-	assert.Contains(buf.String(), fmt.Sprintf("Successfully Deleted 4 files from %v", dirToDelete))
-	buf.Reset()
-
-	// File, not a directory
-	args = []string{"bcda", "delete-dir-contents", "--dirToDelete", "../shared_files/cclf/T.A0001.ACO.ZC8Y18.D181120.T1000009"}
-	err = s.testApp.Run(args)
-	assert.NotNil(err)
-	assert.NotContains(buf.String(), "Successfully Deleted")
-	buf.Reset()
-
-	os.Setenv("TESTDELETEDIRECTORY", "NOT/A/REAL/DIRECTORY")
-	args = []string{"bcda", "delete-dir-contents", "--envvar", "TESTDELETEDIRECTORY"}
-	err = s.testApp.Run(args)
-	assert.NotNil(err)
-	assert.NotContains(buf.String(), "Successfully Deleted")
-	buf.Reset()
-
-}
-
-func (s *CLITestSuite) makeDirToDelete(filePath string) {
-	assert := assert.New(s.T())
-	dirToDelete := filePath
-	err := os.Mkdir(dirToDelete, os.ModePerm)
-	assert.Nil(err)
-
-	_, err = os.Create(filepath.Join(dirToDelete, "deleteMe1.txt"))
-	assert.Nil(err)
-	_, err = os.Create(filepath.Join(dirToDelete, "deleteMe2.txt"))
-	assert.Nil(err)
-	_, err = os.Create(filepath.Join(dirToDelete, "deleteMe3.txt"))
-	assert.Nil(err)
-	_, err = os.Create(filepath.Join(dirToDelete, "deleteMe4.txt"))
-	assert.Nil(err)
-}
-
-func (s *CLITestSuite) setPendingDeletionDir() {
-	err := os.Setenv("PENDING_DELETION_DIR", "/go/src/github.com/CMSgov/bcda-app/bcda/pending_delete_dir")
-	if err != nil {
-		s.FailNow("failed to set the PENDING_DELETION_DIR env variable,", err)
-	}
-	cclfDeletion := os.Getenv("PENDING_DELETION_DIR")
-	err = os.MkdirAll(cclfDeletion, 0744)
-	if err != nil {
-		s.FailNow("failed to create the pending deletion directory,", err)
-	}
-}
-
-func (s *CLITestSuite) resetFiles(resetPath string) {
-	err := filepath.Walk(os.Getenv("PENDING_DELETION_DIR"),
-		func(path string, info os.FileInfo, err error) error {
-			if err != nil {
-				s.FailNow("error in walkfunc,", err)
-			}
-
-			if info.IsDir() {
-				return nil
-			}
-			err = os.Rename(path, resetPath+info.Name())
-			if err != nil {
-				s.FailNow("error in moving files,", err)
-			}
-			return nil
-		})
-	if err != nil {
-		s.FailNow("error in walkfunc,", err)
-	}
 }
