+++ bcda/models/models_test.go
@@ -9,8 +9,8 @@ import (
 	"testing"
 	"time"
 
+	"github.com/CMSgov/bcda-app/bcda/constants"
 	"github.com/CMSgov/bcda-app/bcda/database"
-	"github.com/CMSgov/bcda-app/bcda/testConstants"
 	"github.com/jinzhu/gorm"
 	"github.com/pborman/uuid"
 	"github.com/stretchr/testify/assert"
@@ -183,7 +183,7 @@ func (s *ModelsTestSuite) TestGetEnqueJobs() {
 	assert := s.Assert()
 
 	j := Job{
-		ACOID:      uuid.Parse(testConstants.DEVACOUUID),
+		ACOID:      uuid.Parse(constants.DEVACOUUID),
 		UserID:     uuid.Parse("6baf8254-2e8a-4808-b11d-0fa00c527d2e"),
 		RequestURL: "/api/v1/Patient/$export",
 		Status:     "Pending",
@@ -204,7 +204,7 @@ func (s *ModelsTestSuite) TestGetEnqueJobs() {
 			s.T().Error(err)
 		}
 		assert.Equal(int(j.ID), jobArgs.ID)
-		assert.Equal(testConstants.DEVACOUUID, jobArgs.ACOID)
+		assert.Equal(constants.DEVACOUUID, jobArgs.ACOID)
 		assert.Equal("6baf8254-2e8a-4808-b11d-0fa00c527d2e", jobArgs.UserID)
 		assert.Equal("Patient", jobArgs.ResourceType)
 		assert.Equal(true, jobArgs.Encrypt)
@@ -212,7 +212,7 @@ func (s *ModelsTestSuite) TestGetEnqueJobs() {
 	}
 
 	j = Job{
-		ACOID:      uuid.Parse(testConstants.DEVACOUUID),
+		ACOID:      uuid.Parse(constants.DEVACOUUID),
 		UserID:     uuid.Parse("6baf8254-2e8a-4808-b11d-0fa00c527d2e"),
 		RequestURL: "/api/v1/ExplanationOfBenefit/$export",
 		Status:     "Pending",
@@ -244,7 +244,7 @@ func (s *ModelsTestSuite) TestGetEnqueJobs() {
 func (s *ModelsTestSuite) TestGetBeneficiaryIDs() {
 	assert := s.Assert()
 	var aco, smallACO, mediumACO, largeACO ACO
-	acoUUID := uuid.Parse(testConstants.DEVACOUUID)
+	acoUUID := uuid.Parse(constants.DEVACOUUID)
 
 	err := s.db.Find(&aco, "UUID = ?", acoUUID).Error
 	assert.Nil(err)
@@ -254,7 +254,7 @@ func (s *ModelsTestSuite) TestGetBeneficiaryIDs() {
 	assert.Equal(50, len(beneficiaryIDs))
 
 	// small ACO has 10 benes
-	acoUUID = uuid.Parse(testConstants.SMALLACOUUID)
+	acoUUID = uuid.Parse(constants.SMALLACOUUID)
 	err = s.db.Debug().Find(&smallACO, "UUID = ?", acoUUID).Error
 	assert.Nil(err)
 	beneficiaryIDs, err = smallACO.GetBeneficiaryIDs()
@@ -263,7 +263,7 @@ func (s *ModelsTestSuite) TestGetBeneficiaryIDs() {
 	assert.Equal(10, len(beneficiaryIDs))
 
 	// Medium ACO has 25 benes
-	acoUUID = uuid.Parse(testConstants.MEDIUMACOUUID)
+	acoUUID = uuid.Parse(constants.MEDIUMACOUUID)
 	err = s.db.Find(&mediumACO, "UUID = ?", acoUUID).Error
 	assert.Nil(err)
 	beneficiaryIDs, err = mediumACO.GetBeneficiaryIDs()
@@ -272,7 +272,7 @@ func (s *ModelsTestSuite) TestGetBeneficiaryIDs() {
 	assert.Equal(25, len(beneficiaryIDs))
 
 	// Large ACO has 100 benes
-	acoUUID = uuid.Parse(testConstants.LARGEACOUUID)
+	acoUUID = uuid.Parse(constants.LARGEACOUUID)
 	err = s.db.Find(&largeACO, "UUID = ?", acoUUID).Error
 	assert.Nil(err)
 	beneficiaryIDs, err = largeACO.GetBeneficiaryIDs()
