+++ bcda/testUtils/utils.go
@@ -6,6 +6,10 @@ import (
 	"fmt"
 	"log"
 	"os"
+	"path/filepath"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/suite"
 )
 
 // PrintSeparator prints a line of stars to stdout
@@ -73,8 +77,8 @@ func SetAndRestoreEnvKey(key, value string) func() {
 // by unit tests of the API. Should be called by any API test that uses the auth backend. Returns a function
 // that will restore the original values, suitable for use with defer.
 func SetUnitTestKeysForAuth() func() {
-	private := SetAndRestoreEnvKey("JWT_PRIVATE_KEY_FILE", "../shared_files/api_unit_test_auth_private.pem")
-	public := SetAndRestoreEnvKey("JWT_PUBLIC_KEY_FILE", "../shared_files/api_unit_test_auth_public.pem")
+	private := SetAndRestoreEnvKey("JWT_PRIVATE_KEY_FILE", "../../shared_files/api_unit_test_auth_private.pem")
+	public := SetAndRestoreEnvKey("JWT_PUBLIC_KEY_FILE", "../../shared_files/api_unit_test_auth_public.pem")
 
 	return func() {
 		private()
@@ -84,4 +88,53 @@ func SetUnitTestKeysForAuth() func() {
 	// these paths only work because we assume that the binary location doesn't change from test to test or env to env
 	// the issue, and a path to a more robust solution, is well described in
 	// https://stackoverflow.com/questions/45579312/loading-a-needed-file-relative-vs-absolute-paths
+}
+
+func MakeDirToDelete(s suite.Suite, filePath string) {
+	assert := assert.New(s.T())
+	dirToDelete := filePath
+	err := os.Mkdir(dirToDelete, os.ModePerm)
+	assert.Nil(err)
+
+	_, err = os.Create(filepath.Join(dirToDelete, "deleteMe1.txt"))
+	assert.Nil(err)
+	_, err = os.Create(filepath.Join(dirToDelete, "deleteMe2.txt"))
+	assert.Nil(err)
+	_, err = os.Create(filepath.Join(dirToDelete, "deleteMe3.txt"))
+	assert.Nil(err)
+	_, err = os.Create(filepath.Join(dirToDelete, "deleteMe4.txt"))
+	assert.Nil(err)
+}
+
+func SetPendingDeletionDir(s suite.Suite) {
+	err := os.Setenv("PENDING_DELETION_DIR", "/go/src/github.com/CMSgov/bcda-app/bcda/pending_delete_dir")
+	if err != nil {
+		s.FailNow("failed to set the PENDING_DELETION_DIR env variable,", err)
+	}
+	cclfDeletion := os.Getenv("PENDING_DELETION_DIR")
+	err = os.MkdirAll(cclfDeletion, 0744)
+	if err != nil {
+		s.FailNow("failed to create the pending deletion directory,", err)
+	}
+}
+
+func ResetFiles(s suite.Suite, resetPath string) {
+	err := filepath.Walk(os.Getenv("PENDING_DELETION_DIR"),
+		func(path string, info os.FileInfo, err error) error {
+			if err != nil {
+				s.FailNow("error in walkfunc,", err)
+			}
+
+			if info.IsDir() {
+				return nil
+			}
+			err = os.Rename(path, resetPath+info.Name())
+			if err != nil {
+				s.FailNow("error in moving files,", err)
+			}
+			return nil
+		})
+	if err != nil {
+		s.FailNow("error in walkfunc,", err)
+	}
 }
