+++ bcda/web/middleware.go
@@ -1,4 +1,4 @@
-package main
+package web
 
 import (
 	"github.com/CMSgov/bcda-app/bcda/responseutils"
