+++ bcda/web/router.go
@@ -1,4 +1,4 @@
-package main
+package web
 
 import (
 	"net/http"
@@ -16,8 +16,8 @@ func NewAPIRouter() http.Handler {
 	m := monitoring.GetMonitor()
 	r.Use(auth.ParseToken, logging.NewStructuredLogger(), HSTSHeader, ConnectionClose)
 	// Serve up the swagger ui folder
-	FileServer(r, "/api/v1/swagger", http.Dir("./swaggerui"))
-	FileServer(r, "/", http.Dir("./_site"))
+	FileServer(r, "/api/v1/swagger", http.Dir("../swaggerui"))
+	FileServer(r, "/", http.Dir("../_site"))
 	r.Route("/api/v1", func(r chi.Router) {
 		r.With(auth.RequireTokenAuth, ValidateBulkRequestHeaders).Get(m.WrapHandler("/ExplanationOfBenefit/$export", bulkEOBRequest))
 		if os.Getenv("ENABLE_PATIENT_EXPORT") == "true" {
